+++ clang/lib/Sema/SemaDecl.cpp
@@ -15440,6 +15440,8 @@ void Sema::FinalizeDeclaration(Decl *ThisDecl) {
   CheckInvalidBuiltinCountedByRef(VD->getInit(),
                                   BuiltinCountedByRefKind::Initializer);
 
+  BoundsSafetyCheckCountedByFAMInStaticStorage(VD);
+
   checkAttributesAfterMerging(*this, *VD);
 
   if (VD->isStaticLocal())